Clean up of packa
