+++ Sources/Laka/Subcommands/Images/TextureGenerating.swift
@@ -10,6 +10,7 @@ import Malm
 import Guld
 import Packa
 import Combine
+import Util
 
 protocol TextureGenerating {
     var inDataURL: URL { get }
