+++ Sources/Packa/FittedItems.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Alexander Cyon on 2021-10-09.
+//
+
+import Foundation
+
+public struct FittedItems<Content: Packable> {
+    public let packed: [Packed<Content>]
+    public let canvasSize: CGSize
+}
