+++ Sources/Packa/Packable.swift
@@ -1,0 +1,23 @@
+//
+//  File.swift
+//  
+//
+//  Created by Alexander Cyon on 2021-10-09.
+//
+
+import Foundation
+
+public protocol Packable: Identifiable {
+    var width: CGFloat { get }
+    var height: CGFloat { get }
+}
+
+extension Packable {
+    
+    var size: CGSize { .init(width: width, height: height) }
+    var maxSide: CGFloat { max(width, height) }
+    
+    var area: CGFloat {
+        width * height
+    }
+}
